requirements.txt
cloudify-plugins-common==3.4.0
cloudify-dsl-parser==3.4.0
arundeckrun==0.2.3
requests==1.2.3
